ENH: Search for Python lib. in bin/ sub-dir as well
The Python library may be located in the prefix directory (WinPython),
or in the lib/ sub-directory, or in the bin/ sub-directory (MSYS2).

Search in all those places, but start from a set of root directories and
add sub-directories systematically instead of enumerating each and every
combination directly in a list.