BUG: Handle Python built with GCC when comp. ext.
If the Python interpreter is built with GCC, the version string will of
course not contain a reference to the MSVC runtime. Instead of
attempting to build an import library for yet another runtime, we
assume that the compiler used to build the extension is now the same
as was used to build Python, and that it will pass the right runtime to
the linker without any further command-line options.