BUG: Correct signature of inherited method
The msvccompiler module is made in the same mold as the msvc9compiler
module; saving the two environment variables around the call to
initialize. However, the signaure in the msvccompiler module's version
of the MSVCCompiler class, there is no plat_name argument to the method.

Since this is not used anyway but just passed along, it can simply be
removed.